First upload double linked_list
In this implementation  i have implemented the double linked_list improving the transversing efficiency. Instead of using a while loop to get the last node, i created  a pointer to the keep track of the last node whenever any insertion or deletion is made.  This improves the efficiency for transversing to get the last node.